Hand Class
Hand class was created.
created a function for dealers cards.
The calculate function returns the value of the card, it also asks if a card is an 'Ace'.
An if statement to determined if the 'Ace' is a 1 or and 11.
is_blackjack function checks if a player has blackjack to end the round.